+++ src/screens/Dashboard.js
@@ -5,7 +5,6 @@ import React, { useContext } from 'react';
 import ProductCard from '../components/ProductCard';
 import CartContext from '../context/CartContext';
 
-
 const useStyles = makeStyles(theme => ({
     root: {
       flexGrow: 1,
@@ -29,21 +28,10 @@ const Dashboard = (props) => {
         {
         title: 'Product2',
         description: 'this is product discription. this is product discriptionthis is product discription'
-        },
-        {
-        title: 'Product3',
-        description: 'this is product discription. this is product discriptionthis is product discription'
-        },
-        {
-        title: 'Product4',
-        description: 'this is product discription. this is product discriptionthis is product discription'
         }
     ];
 
-    const products = allItems.map(it=>
-        (
-            {...it, count: 1}
-        ));
+    const products = allItems.map(it=>({...it, count: 1}));
 
     const handleAddProduct = (productToBeAdded) => {
         let tempCart = cart.slice();
@@ -62,37 +50,37 @@ const Dashboard = (props) => {
         setCart(tempCart);
     }
 
-    // const handleRemoveItem = (item) => {
-    //     let tempCart = cart.slice();
-    //     if(item.count === 0){
-    //         tempCart.splice(tempCart.findIndex(it => it.title===item.title), 1);
-    //     }
-    //     else {
-    //         item.count--;
-    //     }
-    //     setCart(tempCart);
-    // }
+    const handleRemoveItem = (item) => {
+        let tempCart = cart.slice();
+        if(item.count === 0){
+            tempCart.splice(tempCart.findIndex(it => it.title===item.title), 1);
+        }
+        else {
+            item.count--;
+        }
+        setCart(tempCart);
+    }
 
     return (
         <div className={classes.root}>
             <Grid container spacing={5}>
                 <Grid item xs={6}>
                     <Paper className={classes.paper}>
-                        <h1>Products</h1>
+                        <h2>Products</h2>
                         {products.map((item)=>(
                             <ProductCard key={item.title} item={item} addItem={handleAddProduct}/>
                         ))}
                     </Paper>
                 </Grid>
-                {/* <Grid item xs={6}>
+                <Grid item xs={6}>
                     <Paper className={classes.paper}>
                         <h2>Cart</h2>
                         {cart.map((item)=>(
                             <ProductCard key={item.title} item={item} removeItem={handleRemoveItem}/>
                         ))}
                     </Paper>
-                </Grid> */}
-            </Grid>          
+                </Grid>
+            </Grid>
         </div>
     )
 }
